backend/requirements.txt
# Backend Requirements
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.0
bson==0.5.10